requirements.txt
acme==2.11.0
APScheduler==3.10.4
asgiref==3.8.1
Brotli==1.1.0
certbot==2.11.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ConfigArgParse==1.7
configobj==5.0.9
cryptography==43.0.1
distro==1.9.0
dj-database-url==2.2.0
Django==5.1
django-apscheduler==0.6.2
django-bootstrap==0.2.4
django-environ==0.11.2
djangorestframework==3.15.2
greenlet==3.0.3
gunicorn==23.0.0
idna==3.10
josepy==1.14.0
packaging==24.1
parsedatetime==2.6
psycopg2==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
pyRFC3339==1.1
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
whitenoise==6.7.0